Meme Generator Project/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
Flask==1.1.4
idna==3.3
importlib-metadata==4.11.1
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.7.1
MarkupSafe==1.1.1
numpy==1.18.0
pandas==0.24.0
Pillow==7.2.0
pkg-resources==0.0.0
py4j==0.10.7
pyspark==2.4.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
requests==2.27.1
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==1.0.1
zipp==3.7.0
